+++ CHANGELOG.md
@@ -14,20 +14,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Entities will now have private setters, a static `Create` method and an `Update` method to promote a DDD workflow
   * This includes base entity and the dbcontext setters for the auditable fields
 * A huge permissions overhaul
-  * SuperAdmin role is added by default and will automatically be given all permissions. 
+  * `SuperAdmin` role is added by default and will automatically be given all permissions. 
   * A `User` role with also be created. This role will not have any permissions assigned to it by default and can be removed if you would like **as long as you add another role in addition to `SuperAdmin`**. This will allow for integration tests of the `UserPolicyHandler` to be fully scoped.
   * A permission will be added for each feature with a default name of `CanFEATURENAME`, but can be overridden if desired.
   * Role-permission mappings outside of this need tobe done manually 
   * You will be responsible for managing permissions outside of super admin
   * Policies are no longer part of the craftsman api, but a single policy will be added with each api boundary to swagger to dictate access to that boundary. It has a default of a snake case of `ProjectName`, but can be overridden
+    * **Call out in docs**. Also docs note that scope added in boundary for access should match the scope on your auth server, 
     * If using the built in auth server, this will be added for you. if not, make sure it mataches an appropriate scope in your auth server for this api boundary
 
-  * **features now have a prop for `IsProtected` that, if true, will add an authorization attribute to your endpoint with the `PolicyName` and add tests that check for access**
+  * **Features now have a prop for `IsProtected` that, if true, will add an authorization attribute to your endpoint with the `PolicyName` and add tests that check for access**
   * Integration tests for `UserPolicyHandler`
   * Unit tests for `RolePermission`
   * `Policies` prop removed from Feature
   * Added `SetUserRole` and `SetUserRoles` methods to integreation tests' `TestFixture` for easy role management in integration tests
   * Functional tests auth helper method now sets `role` instead of `scope`
+  * Alice is a `SuperAdmin` and bob is a `User`
 
 
 ### Updated